refactor(sql-editor): redefine sql editor component structure
