+++ frontend/src/views/SqlEditor/AsidePanel/DatabasesTree.vue
@@ -1,0 +1,1 @@
+<template>Databases Tree</template>
