+++ frontend/src/views/SqlEditor/AsidePanel/TableSchema.vue
@@ -1,0 +1,1 @@
+<template>Table Schema</template>
