+++ frontend/src/views/SqlEditor/EditorPanel/EditorActions.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="sqleditor-editor-actions">Editor Actions</div>
+</template>
+
+<style scoped>
+.sqleditor-editor-actions {
+  @apply w-full;
+}
+</style>
