+++ frontend/src/views/SqlEditor/EditorPanel/EditorPanel.vue
@@ -1,0 +1,10 @@
+<template>
+  <EditorActions />
+
+  <QueryEditor />
+</template>
+
+<script lang="ts" setup>
+import EditorActions from "./EditorActions.vue";
+import QueryEditor from "./QueryEditor.vue";
+</script>
