+++ frontend/src/views/SqlEditor/EditorPanel/QueryEditor.vue
@@ -1,0 +1,3 @@
+<template>Query Editor</template>
+
+<script lang="ts" setup></script>
