+++ frontend/src/views/SqlEditor/TablePanel/HistoryView.vue
@@ -1,0 +1,1 @@
+<template>History View</template>
