+++ frontend/src/views/SqlEditor/TablePanel/TableView.vue
@@ -1,0 +1,1 @@
+<template>Table View</template>
