--- frontend/src/views/SqlEditor/TableView.vue
@@ -1,8 +1,0 @@
-<template>
-  <Splitpanes class="default-theme">
-    <Pane size="100"> Table View </Pane>
-    <Pane size="0"> History </Pane>
-  </Splitpanes>
-</template>
-
-<script lang="ts" setup></script>
