--- frontend/src/views/SqlEditor/aside-panel/index.vue
@@ -1,8 +1,0 @@
-<template>
-  <Splitpanes horizontal class="default-theme">
-    <Pane size="100">Databases Tree</Pane>
-    <Pane size="0" max-size="40">Table Schema</Pane>
-  </Splitpanes>
-</template>
-
-<script lang="ts" setup></script>
